Mouse buttons, weapon ranges/atk, database
Mouse handler supports left, middle, right clicks
Fixed database error with Javelin range "1~1" -> "1~2"
Unit attack and ranges now determined by weapon
button and bar lists changed to HashMaps for functionality improvement

+++ src/gui/Bar.java
@@ -29,7 +29,7 @@ public class Bar implements guiObj
         this.iD = iD;
         this.barColorF = new Color(255, 0, 0, 255);
         this.barColorB = new Color(255, 255, 255, 255);
-        this.value = this.dValue = (float) 1;
+        this.value = this.dValue = 1f;
     }
     
     public Bar(int x, int y, int w, int h, int iD, Color barColorF)
@@ -41,7 +41,7 @@ public class Bar implements guiObj
         this.iD = iD;
         this.barColorF = barColorF;
         this.barColorB = new Color(255, 255, 255, 255);
-        this.value = this.dValue = (float) 1;
+        this.value = this.dValue = 1f;
     }
     
     public void tick()
